Added triggerUpdate in useSaveTimeRecords and updated test setup to support functions in the store
